requirements.txt
aiogram~=3.22.0
aiohttp~=3.12.15
sqlalchemy~=2.0.0
python-dotenv~=1.0.0
pytz~=2023.0